+++ packages/frontend/src/app/modules/common/common-layout.module.ts
@@ -84,30 +84,6 @@ import { LoopStepMentionItemComponent } from './components/form-controls/interpo
 import { CustomPathMentionDialogComponent } from './components/form-controls/interpolating-text-form-control/mentions-list/custom-path-mention-dialog/custom-path-mention-dialog.component';
 import { WarningBoxComponent } from './components/warning-box/warning-box.component';
 import { PieceTriggerMentionItemComponent } from './components/form-controls/interpolating-text-form-control/mentions-list/piece-trigger-mention-item/piece-trigger-mention-item.component';
-import { MarkdownModule, MarkedOptions, MarkedRenderer } from 'ngx-markdown';
-import { MarkdownComponent } from './components/markdown/markdown.component';
-
-export function markedOptionsFactory(): MarkedOptions {
-  const renderer = new MarkedRenderer();
-  const linkRenderer = renderer.link;
-
-  renderer.link = (href, title, text) => {
-    const html = linkRenderer.call(renderer, href, title, text);
-    return html.replace(
-      /^<a /,
-      '<a role="link" tabindex="0" target="_blank" rel="nofollow noopener noreferrer" '
-    );
-  };
-
-  return {
-    renderer,
-    gfm: true,
-    breaks: false,
-    pedantic: false,
-    smartLists: true,
-    smartypants: false,
-  };
-}
 
 export const materialTooltipDefaults: MatTooltipDefaultOptions = {
   showDelay: 0,
@@ -159,7 +135,6 @@ export const materialTooltipDefaults: MatTooltipDefaultOptions = {
     HotspotComponent,
     LoopStepMentionItemComponent,
     CustomPathMentionDialogComponent,
-    MarkdownComponent,
     WarningBoxComponent,
     PieceTriggerMentionItemComponent,
   ],
@@ -195,12 +170,6 @@ export const materialTooltipDefaults: MatTooltipDefaultOptions = {
     MatIconModule,
     MatDividerModule,
     MatTreeModule,
-    MarkdownModule.forRoot({
-      markedOptions: {
-        provide: MarkedOptions,
-        useFactory: markedOptionsFactory,
-      },
-    }),
     MatButtonToggleModule,
   ],
   exports: [
@@ -221,7 +190,6 @@ export const materialTooltipDefaults: MatTooltipDefaultOptions = {
     MatButtonModule,
     OAuth2ConnectControlComponent,
     DictionaryFormControlComponent,
-    MarkdownComponent,
     ConfigsFormComponent,
     CodeArtifactFormControlComponent,
     MatTooltipModule,
