--- packages/frontend/src/app/modules/common/components/loading-skeleton/loading-skeleton.component.html
@@ -1,25 +1,0 @@
-<div class="row mb-4 mt-3">
-  <div class="col-5">
-    <ngx-skeleton-loader  count="1"
-                          appearance="line"></ngx-skeleton-loader>
-  </div>
-</div>
-
-<div class="row mb-2"  *ngFor="let row of emptyArray">
-  <div class="col-4">
-    <ngx-skeleton-loader  count="1"
-                          appearance="line"></ngx-skeleton-loader>
-  </div>
-
-  <div class="col-1">
-
-  </div>
-  <div class="col-3">
-    <ngx-skeleton-loader  count="1"
-                          appearance="line"></ngx-skeleton-loader>
-  </div>
-  <div class="col-3">
-    <ngx-skeleton-loader  count="1"
-                          appearance="line"></ngx-skeleton-loader>
-  </div>
-</div>
