--- packages/frontend/src/app/modules/common/components/loading-skeleton/loading-skeleton.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-loading-skeleton',
-  templateUrl: './loading-skeleton.component.html',
-  styleUrls: ['./loading-skeleton.component.css'],
-})
-export class LoadingSkeletonComponent {
-  emptyArray = [{}, {}, {}];
-}
