+++ packages/frontend/src/app/modules/flow-builder/page/flow-builder/flow-right-sidebar/edit-step-sidebar/edit-step-accordion/input-forms/piece-input-forms/custom-auth-connection-dialog/custom-auth-connection-dialog.component.html
@@ -4,7 +4,7 @@
 </app-dialog-title-template>
 
 <mat-dialog-content>
-    <app-markdown *ngIf="dialogData.pieceAuthConfig.description" [data]="dialogData.pieceAuthConfig.description"></app-markdown>
+    <ap-markdown *ngIf="dialogData.pieceAuthConfig.description" [data]="dialogData.pieceAuthConfig.description"></ap-markdown>
     <form class="ap-pt-4 ap-flex ap-flex-col ap-gap-2 ap-max-w-[720px] ap-min-w-[420px]" [formGroup]="settingsForm" (submit)="submit()">
         <mat-form-field class="ap-w-full" appearance="outline">
             <mat-label>Name</mat-label>
