+++ packages/frontend/src/app/modules/flow-builder/page/flow-builder/flow-right-sidebar/edit-step-sidebar/edit-step-accordion/input-forms/piece-input-forms/secret-text-connection-dialog/secret-text-connection-dialog.component.html
@@ -3,7 +3,7 @@
 </app-dialog-title-template>
 
 <mat-dialog-content>
-  <app-markdown *ngIf="dialogData.description" [data]="dialogData.description"></app-markdown>
+  <ap-markdown *ngIf="dialogData.description" [data]="dialogData.description"></ap-markdown>
   <form class="ap-pt-4 ap-flex ap-flex-col ap-gap-2 ap-max-w-[720px] ap-min-w-[420px]" [formGroup]="settingsForm" (submit)="submit()">
     <mat-form-field class="ap-w-full" appearance="outline">
       <mat-label>Name</mat-label>
