+++ packages/ui/common/README.md
@@ -1,0 +1,7 @@
+# ui-common
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test ui-common` to execute the unit tests.
