+++ tsconfig.base.json
@@ -125,6 +125,7 @@
       "@activepieces/piece-zoom": ["packages/pieces/zoom/src/index.ts"],
       "@activepieces/pieces-apps": ["packages/pieces/apps/src/index.ts"],
       "@activepieces/shared": ["packages/shared/src/index.ts"],
+      "@activepieces/ui/common": ["packages/ui/common/src/index.ts"],
       "@backend/*": ["packages/backend/src/app/*"],
       "@ee/*": ["packages/ee/*"],
       "@frontend/*": ["packages/frontend/src/app/*"],
